Show the summary proportion in a widget

A ratewidget displays a single statistic derived from a linked table, where the numerator and denominator columns are specific. This is best used with ] the crosstalkpackage, to allow dynamic filtering of the underlying data.

Usage

ratewidget(
  data,
  numerator,
  denominator,
  multiplier = 100,
  digits = 0,
  width = NULL,
  height = NULL,
  elementId = NULL
)

Arguments

data

A data frame or data table containing the data to summarize, with at least two numeric columns: one for the numerator and one for the denominator. Normally, this should be an instance of crosstalk::SharedData.

numerator

The name of the column in data to use as the numerator.

denominator

The name of the column in data to use as the denominator.

multiplier

A numeric value to multiply the proportion by (default is 100, to express as a percentage).

digits

The number of decimal places to round the result to (default is 0).

width

The width of the widget (optional).

height

The height of the widget (optional).

elementId

An optional element ID for the widget.